+++ project/Source/FallingTrap.cpp
@@ -7,8 +7,8 @@
 FallingTrap::FallingTrap(int sx, int sy)
 {
 	hImage = LoadGraph("data/image/New Blo.png");
-	x = static_cast<float>(sx);
-	y = static_cast<float>(sy);
+	x = sx;
+	y = sy;
 
 	velocityY = 0.0f;
 	isFalling = false;
